Notes list now available on notes screen
notes retrieved from server and added to model.
noteList returns size (for iterating through).
added displayNotes method in view that takes a noteList.
main calls view.displayNotes(model.getNotes) but this should be in the
switch case in the listener for when notes is clicked; as the listener
is in the view, which has no knowledge of the model, it cannot call
model.getNotes.
tried to write a workaround; model passes to view, passes to notes
screen on button press but the list somehow becomes empty.
realised this was because button listeners were creating a new view and
then applying the changes to this view which, naturally, did not have
the same values as the old view.
changed the view package so that each screen takes this view, and each
button listener takes a view.